fix data dashboard layout
