+++ src/pages/admin/hampersPage.jsx
@@ -1,6 +1,6 @@
 const HampersPage = () => {
     return (
-        <div className='w-screen p-4 overflow-y-auto'>
+        <div className='w-screen min-h-screen p-4 overflow-y-auto'>
             <div className="flex items-center">
                 <h1 className="text-4xl text-[#d08854] font-semibold">Data Hampers</h1>
                 <div className="divider divider-horizontal m-1"></div>
@@ -8,23 +8,22 @@ const HampersPage = () => {
                     Manajemen Hampers Atma Kitchen
                 </p>
             </div>
-            <div className="card w-full min-h-screen bg-white mt-2">
-                <div className="card-body p-4">
+            <div className="card w-full h-fit bg-white mt-4">
+                <div className="card-body h-full p-4">
                     <div className='flex justify-between'>
                         <div className="flex items-center">
-                            <label className="label">Cari Hampers: </label>
-                            <input type="text" className='input bg-slate-200' />
+                            <input type="text" placeholder="Cari Hampers" className='input bg-slate-100' />
                         </div>
                         <div className="space-x-1">
                             <button className='btn btn-outline bg-[#d08854] text-white'>Tambah Hampers</button>
                         </div>
                     </div>
                     <div className="divider m-1"></div>
+                    <div className='mt-2'>
+                        {/* <DataTable columns={columns} data={records} fixedHeader pagination className='mt-2'>
+                        </DataTable> */}
+                    </div>
                 </div>
-            </div>
-            <div className='mt-2'>
-                {/* <DataTable columns={columns} data={records} fixedHeader pagination className='mt-2'>
-                </DataTable> */}
             </div>
         </div>
     )
